% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_join_cols.R
\name{get_join_cols}
\alias{get_join_cols}
\title{cols that would be joined and their origin}
\usage{
get_join_cols(data_x, data_y)
}
\arguments{
\item{data_x}{data x}

\item{data_y}{data y (will be joined on x)}
}
\value{
df with all columns of a join and their origin
}
\description{
cols that would be joined and their origin
}
\examples{
## get_join_cols(df.x, df.y)
}
