% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_join_count.R
\name{get_join_count}
\alias{get_join_count}
\title{Count Join ID in a dataset}
\usage{
get_join_count(data_x, ID_x)
}
\arguments{
\item{data_x}{data}

\item{ID_x}{Join ID in dataset x}
}
\value{
df with how many times an ID appears in a dataset
}
\description{
Count Join ID in a dataset
}
\examples{
##get_join_count(df.x, df.x$ID)
}
