% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_join_rowindex.R
\name{get_join_rowindex}
\alias{get_join_rowindex}
\title{col indicating join partner in y}
\usage{
get_join_rowindex(data_x, ID_x, ID_y)
}
\arguments{
\item{data_x}{data x that the join will be based on}

\item{ID_x}{Join ID in dataset x}

\item{ID_y}{Join ID in dataset y}
}
\value{
df x with new column join_index (0 = FALSE)
}
\description{
col indicating join partner in y
}
\examples{
##get_join_rowindex(df.x, df.x$ID, df.y$ID)
}
