% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_join_rows_n.R
\name{get_join_rows_n}
\alias{get_join_rows_n}
\title{count of rows with join}
\usage{
get_join_rows_n(data_x, ID_x, ID_y)
}
\arguments{
\item{data_x}{data x that the join will be based on}

\item{ID_x}{Join ID in dataset x}

\item{ID_y}{Join ID in dataset y}
}
\value{
dataframe of n rows that have a join match in x
}
\description{
count of rows with join
}
\examples{
##get_join_rows_n(df.x, df.x$ID, df.y$ID)
}
